v0.9.0
v0.9.0 introduces new interfaces for Axona, some general schema fixes for lfp interface metadata, as well as GIN compatible tests for certain ecephys interfaces. It also introduces new methods of data chunking for write_recording as well as a change to the keyword arguments of all routers to that function.